feat(security): Add CSRF token protection on every form (#64)
Closes #62

To enhance security of the service, CSRF protection is added to every HTML form of the application to prevent such attacks. The protection is simple, every time a user browses a page containing a form, a unique csrf token is injected in a hidden input. This token is valid only once and is required for data submission. This will prevent programs from doing POST requests without passing by the form.